Move uploadFileToGoogleDrive to media-export service
